backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
python-jose[cryptography]
bcrypt
pydantic
pydantic-settings
mutagen
python-multipart

